upgrade project according to 'PEP8'
