+++ gb_client-server_apps/lesson_2_7/client/del_contact.py
@@ -1,0 +1,53 @@
+"""Модуль удаления контакта."""
+
+import logging
+from PyQt5.QtWidgets import QDialog, QLabel, QComboBox, QPushButton, \
+    QApplication
+from PyQt5.QtCore import Qt
+
+
+LOG = logging.getLogger('app.client')
+
+
+class DelContactDialog(QDialog):
+    """
+    Класс - диалог удаления контакта.
+
+    Предлагает текущий список контактов, не имеет обработчиков для действий.
+    """
+
+    def __init__(self, database):
+        super().__init__()
+        self.database = database
+
+        self.setFixedSize(350, 120)
+        self.setWindowTitle('Выберите контакт для удаления:')
+        self.setAttribute(Qt.WA_DeleteOnClose)
+        self.setModal(True)
+
+        self.selector_label = QLabel('Выберите контакт для удаления:', self)
+        self.selector_label.setFixedSize(200, 20)
+        self.selector_label.move(10, 0)
+
+        self.selector = QComboBox(self)
+        self.selector.setFixedSize(200, 20)
+        self.selector.move(10, 30)
+
+        self.btn_ok = QPushButton('Удалить', self)
+        self.btn_ok.setFixedSize(100, 30)
+        self.btn_ok.move(230, 20)
+
+        self.btn_cancel = QPushButton('Отмена', self)
+        self.btn_cancel.setFixedSize(100, 30)
+        self.btn_cancel.move(230, 60)
+        self.btn_cancel.clicked.connect(self.close)
+
+        # заполнитель контактов для удаления
+        self.selector.addItems(sorted(self.database.get_contacts()))
+
+
+if __name__ == '__main__':
+    app = QApplication([])
+    window = DelContactDialog(None)
+    window.show()
+    app.exec_()
