+++ gb_client-server_apps/lesson_2_7/common/errors.py
@@ -1,0 +1,14 @@
+"""Модуль, содержащий шаблоны стандартных ошибок приложения."""
+
+
+class ServerError(Exception):
+    """Класс - исключение, для обработки ошибок сервера.
+
+    При генерации требует строку с описанием ошибки, полученную с сервера.
+    """
+
+    def __init__(self, text):
+        self.text = text
+
+    def __str__(self):
+        return self.text
