+++ gb_client-server_apps/lesson_2_7/common/metaclasses.py
@@ -1,0 +1,84 @@
+"""Модуль, содержащий метаклассы, для контроля классов приложения."""
+
+import dis
+
+
+class ServerMaker(type):
+    """Метакласс проверки данных классов.
+
+    Проверяет, что в результирующем классе нет клиентских вызовов таких как:
+    connect. Также проверяется, что серверный сокет является TCP и работает
+    по IPv4 протоколу.
+    """
+
+    def __init__(cls, clsname, bases, clsdict):
+        # Список методов, которые используются в функциях класса:
+        methods = []
+        # Атрибуты, вызываемые функциями классов
+        attrs = []
+        for func in clsdict:
+            # Пробуем
+            try:
+                ret = dis.get_instructions(clsdict[func])
+                # Если не функция то ловим исключение
+            except TypeError:
+                pass
+            else:
+                # Раз функция разбираем код, получая используемые методы и
+                # атрибуты.
+                for i in ret:
+                    if i.opname == 'LOAD_GLOBAL':
+                        if i.argval not in methods:
+                            methods.append(i.argval)
+                    elif i.opname == 'LOAD_ATTR':
+                        if i.argval not in attrs:
+                            attrs.append(i.argval)
+        # Если обнаружено использование недопустимого метода connect,
+        # генерируем исключение:
+        if 'connect' in methods:
+            raise TypeError(
+                'Использование метода connect недопустимо в серверном классе')
+        # Если сокет не инициализировался константами SOCK_STREAM(TCP)
+        # AF_INET(IPv4), тоже исключение.
+        if not ('SOCK_STREAM' in attrs and 'AF_INET' in attrs):
+            raise TypeError('Некорректная инициализация сокета.')
+        super().__init__(clsname, bases, clsdict)
+
+
+class ClientMaker(type):
+    """Метакласс проверки данных классов.
+    Проверяет, что в результирующем классе нет серверных вызовов таких как:
+    accept, listen. Также проверяется, что сокет не создаётся внутри
+    конструктора класса.
+    """
+
+    def __init__(cls, clsname, bases, clsdict):
+        # Список методов, которые используются в функциях класса:
+        methods = []
+        for func in clsdict:
+            # Пробуем
+            try:
+                ret = dis.get_instructions(clsdict[func])
+                # Если не функция, то ловим исключение
+            except TypeError:
+                pass
+            else:
+                # Раз функция разбираем код, получая используемые методы.
+                for i in ret:
+                    if i.opname == 'LOAD_GLOBAL':
+                        if i.argval not in methods:
+                            methods.append(i.argval)
+        # Если обнаружено использование недопустимого метода accept, listen,
+        # socket бросаем исключение:
+        for command in ('accept', 'listen', 'socket'):
+            if command in methods:
+                raise TypeError(
+                    'В классе обнаружено использование запрещённого метода')
+        # Вызов get_message или send_message из utils считаем корректным
+        # использованием сокетов
+        if 'get_message' in methods or 'send_message' in methods:
+            pass
+        else:
+            raise TypeError(
+                'Отсутствуют вызовы функций, работающих с сокетами.')
+        super().__init__(clsname, bases, clsdict)
