+++ gb_client-server_apps/lesson_2_7/common/utils.py
@@ -1,0 +1,40 @@
+"""Модуль, содержащий общие функции приложения."""
+
+from .decos import log
+from .variables import MAX_PACKAGE_LENGTH, ENCODING
+import json
+import sys
+
+sys.path.append('../')
+
+
+@log
+def get_message(client):
+    """Функция приёма сообщений от удалённых компьютеров.
+
+    Принимает сообщения JSON, декодирует полученное сообщение
+    и проверяет что получен словарь.
+    :param client: сокет для передачи данных.
+    :return: словарь - сообщение.
+    """
+    encoded_response = client.recv(MAX_PACKAGE_LENGTH)
+    json_response = encoded_response.decode(ENCODING)
+    response = json.loads(json_response)
+    if isinstance(response, dict):
+        return response
+    else:
+        raise TypeError
+
+
+@log
+def send_message(sock, message):
+    """Функция отправки словарей через сокет.
+
+    Кодирует словарь в формат JSON и отправляет через сокет.
+    :param sock: сокет для передачи.
+    :param message: словарь для передачи.
+    :return: ничего не возвращает.
+    """
+    js_message = json.dumps(message)
+    encoded_message = js_message.encode(ENCODING)
+    sock.send(encoded_message)
